Move the code using libcap to separate files
So we don't have to link every tool against libcap.
