+++ pdns/dnsdistdist/capabilities.hh
@@ -1,0 +1,1 @@
+../capabilities.hh
