+++ pdns/misc.hh
@@ -593,7 +593,6 @@ double DiffTime(const struct timespec& first, const struct timespec& second);
 double DiffTime(const struct timeval& first, const struct timeval& second);
 uid_t strToUID(const string &str);
 gid_t strToGID(const string &str);
-void dropCapabilities();
 
 unsigned int pdns_stou(const std::string& str, size_t * idx = 0, int base = 10);
