+++ pdns/recursordist/capabilities.cc
@@ -1,0 +1,1 @@
+../capabilities.cc
